requirements.txt
mcp>=1.12.0
pyserial>=3.5
